Segregate common trusted signup APIs
Common signup related APIs between different type of workers are added
to separate file.
